requirements.txt
langchain-openai
langchain
langchain-core
langchain-groq

langgraph
matplotlib
mplfinance
plotly
scikit-learn
python-dotenv
langchain_community
numpy
datetime
yfinance
nltk
streamlit
yfinance
pandas
numpy
plotly
scikit-learn
python-dotenv

langchain
langchain-groq
langchain-community

tavily-python
websocket-client
sec-api beautifulsoup4 requests

nltk
requests
